Add Circuit logic to ai bot
