+++ build.gradle.kts
@@ -32,7 +32,6 @@ import org.jetbrains.intellij.tasks.BuildPluginTask
 import org.jetbrains.intellij.tasks.PatchPluginXmlTask
 import org.jetbrains.kotlin.gradle.dsl.JvmTarget
 import org.jetbrains.kotlin.gradle.dsl.KotlinJvmProjectExtension
-import org.jetbrains.kotlin.gradle.dsl.KotlinVersion.KOTLIN_1_8
 import org.jetbrains.kotlin.gradle.dsl.KotlinVersion.KOTLIN_1_9
 import org.jetbrains.kotlin.gradle.plugin.KotlinBasePlugin
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
@@ -221,7 +220,7 @@ subprojects {
       compilerOptions {
         val kotlinVersion =
           if (isForIntelliJPlugin) {
-            KOTLIN_1_8
+            KOTLIN_1_9
           } else {
             KOTLIN_1_9
           }
