+++ skate-plugin/project-gen/src/jvmMain/kotlin/slack/tooling/aibot/ChatPanel.kt
@@ -15,16 +15,31 @@
  */
 package slack.tooling.aibot
 
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.awt.ComposePanel
+import com.slack.circuit.foundation.Circuit
+import com.slack.circuit.foundation.CircuitContent
 import java.awt.Dimension
 import javax.swing.JComponent
+import slack.tooling.aibot.ChatWindowUi.ChatWindow
 import slack.tooling.projectgen.SlackDesktopTheme
 
 object ChatPanel {
   fun createPanel(): JComponent {
     return ComposePanel().apply {
       preferredSize = Dimension(400, 600)
-      setContent { SlackDesktopTheme { ChatWindow() } }
+      setContent { SlackDesktopTheme { ChatApp() } }
     }
+  }
+
+  @Composable
+  private fun ChatApp() {
+    val circuit =
+      Circuit.Builder()
+        .addPresenter<ChatScreen, ChatScreen.State>(ChatPresenter())
+        .addUi<ChatScreen, ChatScreen.State> { state, modifier -> ChatWindow(state, modifier) }
+        .build()
+
+    CircuitContent(ChatScreen, circuit = circuit)
   }
 }
